Explicitly cast float to integer to avoid warning about implicit conversion

+++ code/botlib/l_precomp.c
@@ -1804,7 +1804,7 @@ int PC_Directive_eval(source_t* source)
     token.whitespace_p = source->scriptstack->script_p;
     token.endwhitespace_p = source->scriptstack->script_p;
     token.linescrossed = 0;
-    sprintf(token.string, "%d", abs(value));
+    sprintf(token.string, "%d", abs((int)value));
     token.type = TT_NUMBER;
     token.subtype = TT_INTEGER | TT_LONG | TT_DECIMAL;
     PC_UnreadSourceToken(source, &token);
